Break crate into modules
