+++ commons/src/errors.rs
@@ -1,0 +1,3 @@
+use std::error::Error;
+
+pub type AsyncError = Box<dyn Error + Send + Sync>;
