+++ database-api/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "database-api"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde_json = "1.0"
+async-trait = "0.1"
+
+[dependencies.serde]
+workspace = true
+features = ["derive"]
+
+[dependencies.tokio]
+workspace = true
+features = ["macros", "rt-multi-thread", "fs", "sync"]
+default-features = false
