+++ discord-api/src/lib.rs
@@ -1,5 +1,7 @@
 mod commands;
 mod webhook;
 
+pub mod config;
+
 pub use commands::Gateway;
 pub use webhook::*;
