+++ twitch-api/src/error.rs
