+++ twitch-api/src/model.rs
